Access Data from Giphy and Return JSON
The Express server now has one route, which handles a
request for data from the Giphy API using the secret key stored
server-side.

The search term is passed via the URL's 'search' query (defaults
to a blank string). For the time being, we've hard-coded the number
of gif records that we want Giphy to return: 3. Our gifs should be
PG-rated.