Add Logo and Some Styles
This is the first run-through on styling the app. We added a logo
that disappears when the user begins to type their query. This leaves
more space to view the gifs that are returned.

There is still some work to do on styling the individual images, as
well as figuring out some general positioning questions.